+++ src/main/java/Model/Observer/SeatViewer.java
@@ -4,15 +4,15 @@ package Model.Observer;
  * Created by Коля on 14.09.2015.
  */
 public class SeatViewer {
-    private int seatId;
+    private long seatId;
     private int seatNum;
 
-    public SeatViewer(int seatId, int seatNum) {
+    public SeatViewer(long seatId, int seatNum) {
         this.seatId = seatId;
         this.seatNum = seatNum;
     }
 
-    public int getSeatId() {
+    public long getSeatId() {
         return seatId;
     }
